exercise: made changes to the materials in the visual and the imported a materials xacro
The colour files for robot model are different for Gazebo and Rviz. The urdf file specifies the colours for the robot in Rviz - this is comprised of a material tag in each of the visual tags for the robot links and a materials xacro file that needs to be imported as well. The gazebo file specifies the colours directly for each link.